pyproject.toml
[project]
name = "gridaibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "openai>=2.21.0",
    "python-dotenv>=1.0.0",
    "python-okx>=0.4.1",
    "websockets>=16.0",
]
[tool.pyright]
reportMissingTypeStubs = false
